feat: added Dockerfile, and added helpers function
