+++ middleware/dummy_script.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-echo "This is a dummy script running in a Docker container!"
-echo "Environment Variable: $DUMMY_VARIABLE"
+# Install python3 requirements
+pip install --no-cache-dir -r requirements.txt
 
-# Keep the container running
-tail -f /dev/null
+# Run the main.py
+python main.py
