+++ middleware/src/api/endpoints.py
@@ -1,6 +1,7 @@
 # api/endpoints.py
 from flask import request, jsonify
 import requests
+from utils.helpers import webscraper, ollama
 
 def register_endpoints(app, get_services, save_service, get_service):
     @app.route('/api/v1/get-specs', methods=['GET'])
@@ -23,22 +24,7 @@ def register_endpoints(app, get_services, save_service, get_service):
 
             return jsonify({'error': 'Servizio non trovato e impossibile ottenere i dati dal componente'}), 404
 
-    def webscraper(service_name):
-        component_url = f'http://localhost:5001/api/v1/get-component'
-        response = requests.get(component_url, params={'name': service_name})
-
-        if response.status_code == 200:
-            return response.json()
-
-        # Altrimenti, restituisci None
-        return None
     
-    def ollama(website):
-        component_url = f'http://localhost:5002/api/v1/get-specs'
-        response = requests.get(component_url, params={'website': website})
-
-        if response.status_code == 200:
-            return response.json()
-
-        # Altrimenti, restituisci None
-        return None
+    @app.route('/api/v1/get-services', methods=['GET'])
+    def get_all_services():
+        return get_services()
